Move all python files into a top level package folder
