+++ snp_oracle/__init__.py
@@ -1,0 +1,5 @@
+import importlib.metadata
+
+__version__ = importlib.metadata.version(__name__ or __package__)
+version_split = __version__.split(".")
+__spec_version__ = (1000 * int(version_split[0])) + (10 * int(version_split[1])) + (1 * int(version_split[2]))
