+++ snp_oracle/base_miner/predict.py
@@ -1,7 +1,3 @@
-# developer: Foundry Digital
-# Copyright © 2023 Foundry Digital
-
-# Import modules that already exist or can be installed using pip
 from datetime import datetime
 
 import numpy as np
@@ -10,9 +6,8 @@ from pytz import timezone
 from sklearn.preprocessing import MinMaxScaler
 from tensorflow.keras.models import load_model
 
-# import custom defined files
-from base_miner.get_data import prep_data, round_down_time, scale_data
-from base_miner.model import create_and_save_base_model_lstm
+from snp_oracle.base_miner.get_data import prep_data, round_down_time, scale_data
+from snp_oracle.base_miner.model import create_and_save_base_model_lstm
 
 
 def predict(timestamp: datetime, scaler: MinMaxScaler, model, type) -> float:
