+++ snp_oracle/predictionnet/__init__.py
@@ -1,0 +1,5 @@
+import json
+
+# Import all submodules.
+from snp_oracle.predictionnet import base, protocol, validator
+
