+++ snp_oracle/predictionnet/base/neuron.py
@@ -1,30 +1,12 @@
-# The MIT License (MIT)
-# Copyright © 2023 Yuma Rao
-
-# Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated
-# documentation files (the “Software”), to deal in the Software without restriction, including without limitation
-# the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software,
-# and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
-
-# The above copyright notice and this permission notice shall be included in all copies or substantial portions of
-# the Software.
-
-# THE SOFTWARE IS PROVIDED “AS IS”, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO
-# THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL
-# THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
-# OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
-# DEALINGS IN THE SOFTWARE.
-
 import copy
 from abc import ABC, abstractmethod
 
 import bittensor as bt
 
-from predictionnet import __spec_version__ as spec_version
+from snp_oracle import __spec_version__ as spec_version
 
-# Sync calls set weights and also resyncs the metagraph.
-from predictionnet.utils.config import add_args, check_config, config
-from predictionnet.utils.misc import ttl_get_block
+from snp_oracle.predictionnet.utils.config import add_args, check_config, config
+from snp_oracle.predictionnet.utils.misc import ttl_get_block
 
 
 class BaseNeuron(ABC):
