+++ snp_oracle/predictionnet/utils/__init__.py
@@ -1,0 +1,1 @@
+from snp_oracle.predictionnet.utils import config, misc, uids
