+++ snp_oracle/predictionnet/validator/__init__.py
@@ -1,0 +1,2 @@
+from snp_oracle.predictionnet.validator.forward import forward
+from snp_oracle.predictionnet.validator.reward import get_rewards
